test(web-builder): move test files into `src`
